naver-blog-agents/blog-image-agent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "blog-image-agent"
version = "1.0.0"
description = "블로그 이미지 자동 수집 및 생성 에이전트"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    "click>=8.1.0",
    "httpx>=0.27.0",
    "pillow>=10.0.0",
    "beautifulsoup4>=4.12.0",
    "google-genai>=0.4.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0.0",
    "aiofiles>=23.2.0",
    "lxml>=5.0.0",
    "requests>=2.31.0",
    "scipy>=1.11.0",
    "typing-extensions>=4.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
]

[project.scripts]
blog-image = "cli.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
